Split upt tests into test_oggex (unit tests) and test_oggex_int (integration tests)
